Fix build warnings

+++ src/algorithms/fastod/attribute_set.cpp
@@ -9,58 +9,58 @@ using namespace algos::fastod;
 
 AttributeSet::AttributeSet() noexcept : set_(), value_(0) {}
 
-AttributeSet::AttributeSet(int attribute) noexcept : set_({attribute}), value_(1 << attribute) {}
+AttributeSet::AttributeSet(size_t attribute) noexcept : set_({attribute}), value_(1 << attribute) {}
 
-AttributeSet::AttributeSet(const std::vector<int>& attributes) noexcept :
+AttributeSet::AttributeSet(const std::vector<size_t>& attributes) noexcept :
     set_(attributes.cbegin(), attributes.cend()),
-    value_(std::accumulate(attributes.cbegin(), attributes.cend(), 0, [](const unsigned long long& acc, const int& curr){ return acc + (1 << curr); })) {}
+    value_(std::accumulate(attributes.cbegin(), attributes.cend(), 0, [](const unsigned long long& acc, const size_t& curr){ return acc + (1 << curr); })) {}
 
-AttributeSet::AttributeSet(const std::set<int>& set) noexcept :
+AttributeSet::AttributeSet(const std::set<size_t>& set) noexcept :
     set_(set),
-    value_(std::accumulate(set.cbegin(), set.cend(), 0, [](const unsigned long long& acc, const int& curr){ return acc + (1 << curr); })) {}
+    value_(std::accumulate(set.cbegin(), set.cend(), 0, [](const unsigned long long& acc, const size_t& curr){ return acc + (1 << curr); })) {}
 
-bool AttributeSet::ContainsAttribute(int attribute) const noexcept {
+bool AttributeSet::ContainsAttribute(size_t attribute) const noexcept {
     return set_.find(attribute) != set_.end();
 }
 
-AttributeSet AttributeSet::AddAttribute(int attribute) const noexcept {
+AttributeSet AttributeSet::AddAttribute(size_t attribute) const noexcept {
     if (ContainsAttribute(attribute)) {
         return *this;
     }
 
-    std::set<int> new_set = set_;
+    std::set<size_t> new_set = set_;
     new_set.insert(attribute);
 
     return AttributeSet(std::move(new_set));
 }
 
-AttributeSet AttributeSet::DeleteAttribute(int attribute) const noexcept {
+AttributeSet AttributeSet::DeleteAttribute(size_t attribute) const noexcept {
     if (!ContainsAttribute(attribute)) {
         return *this;
     }
 
-    std::set<int> new_set = set_;
+    std::set<size_t> new_set = set_;
     new_set.erase(attribute);
 
     return AttributeSet(std::move(new_set));
 }
 
 AttributeSet AttributeSet::Intersect(const AttributeSet& other) const noexcept {
-    std::vector<int> result;
+    std::vector<size_t> result;
     std::set_intersection(set_.begin(), set_.end(), other.set_.begin(), other.set_.end(), std::back_inserter(result));
 
     return AttributeSet(std::move(result));
 }
 
 AttributeSet AttributeSet::Union(const AttributeSet& other) const noexcept {
-    std::vector<int> result;
+    std::vector<size_t> result;
     std::set_union(set_.begin(), set_.end(), other.set_.begin(), other.set_.end(), std::back_inserter(result));
 
     return AttributeSet(std::move(result));
 }
 
 AttributeSet AttributeSet::Difference(const AttributeSet& other) const noexcept {
-    std::vector<int> result;
+    std::vector<size_t> result;
     std::set_difference(set_.begin(), set_.end(), other.set_.begin(), other.set_.end(), std::back_inserter(result));
 
     return AttributeSet(std::move(result));
@@ -77,7 +77,7 @@ std::string AttributeSet::ToString() const noexcept {
 
    bool first = true;
 
-   for (int attribute: set_) {
+   for (size_t attribute: set_) {
         if (first) {
             first = false;
         } else {
@@ -100,11 +100,11 @@ unsigned long long AttributeSet::GetValue() const noexcept {
     return value_;
 }
 
-std::set<int>::iterator AttributeSet::begin() const noexcept {
+std::set<size_t>::iterator AttributeSet::begin() const noexcept {
     return set_.begin();
 }
 
-std::set<int>::iterator AttributeSet::end() const noexcept {
+std::set<size_t>::iterator AttributeSet::end() const noexcept {
     return set_.end();
 }
